+++ internal/gapicgen/generator/config.go
@@ -1979,6 +1979,23 @@ var MicrogenGapicConfigs = []*MicrogenConfig{
 		// GA after 2023/02/12
 		ReleaseLevel: "beta",
 	},
+	{
+		InputDirectoryPath:    "google/maps/mapsplatformdatasets/v1alpha",
+		Pkg:                   "mapsplatformdatasets",
+		ImportPath:            "cloud.google.com/go/maps/mapsplatformdatasets/apiv1alpha",
+		GRPCServiceConfigPath: "mapsplatformdatasets_grpc_service_config.json",
+		ApiServiceConfigPath:  "mapsplatformdatasets_v1alpha.yaml",
+		ReleaseLevel:          "alpha",
+	},
+	{
+		InputDirectoryPath:    "google/cloud/discoveryengine/v1beta",
+		stopGeneration:        true,
+		Pkg:                   "discoveryengine",
+		ImportPath:            "cloud.google.com/go/discoveryengine/apiv1beta",
+		GRPCServiceConfigPath: "discoveryengine_grpc_service_config.json",
+		ApiServiceConfigPath:  "discoveryengine_v1beta.yaml",
+		ReleaseLevel:          "beta",
+	},
 
 	// Non-Cloud APIs
 	{